libc: Correct C11 threads timeouts and return values
The C11 threads wrappers treated the absolute TIME_UTC deadline required
by mtx_timedlock() and cnd_timedwait() as if it were a relative timeout,
so a deadline computed from timespec_get() waited for the whole wall
clock time since the epoch. Add a shared c11_timeout_ms() helper that
converts the deadline against the current time, reports an expired or
malformed deadline separately, and never yields the zero that KOS
primitives interpret as "wait forever".

mtx_timedlock() now tries the lock before consulting the clock, as the
standard permits it to succeed on an already-expired deadline, and
cnd_timedwait() still releases and reacquires the mutex when the
deadline has passed.

thrd_create() also passed a thrd_start_t through a cast to the
incompatible kthread entry signature and truncated the result through a
pointer, which is undefined behavior and loses the value on targets
where int and void * differ. Wrap the callback in a trampoline that
carries the argument and converts the result through intptr_t, matching
thrd_exit() and thrd_join().

(cherry picked from commit 358974c106bcf0d358723ed7b894e7079ccf670c)

+++ kernel/libc/c11/thrd_join.c
@@ -5,6 +5,7 @@
 */
 
 #include <threads.h>
+#include <stdint.h>
 
 int thrd_join(thrd_t thr, int *res) {
     void *rv;
@@ -13,7 +14,7 @@ int thrd_join(thrd_t thr, int *res) {
         return thrd_error;
 
     if(res)
-        *res = (int)rv;
+        *res = (int)(intptr_t)rv;
 
     return thrd_success;
 }
